+++ spark_data_frame_comparer/spark_data_frame_exception_type.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class ExceptionType(Enum):
+    SchemaMismatch = 1
+    RowMismatch = 2
+    DataMismatch = 3
